requirements.txt
requests
beautifulsoup4
lxml
python-jobspy
openpyxl
pandas
schedule
fuzzywuzzy
python-Levenshtein
